robotiq_deoxys_control/deoxys/requirements.txt
pyzmq
PyYAML
pyasyncore
easydict
google-api-python-client
numpy
hidapi
glfw
numba
termcolor
simplejson
pyasynch
pymodbus==2.5.3